Добавлена реализация разбора апк

+++ src/main/java/com/example/HanldeImg/services/image_service/ImageServiceImpl.java
@@ -1,11 +1,7 @@
 package com.example.HanldeImg.services.image_service;
 
-import com.example.HanldeImg.entity.dto.ImageDTO;
-import com.example.HanldeImg.entity.model.Image;
-import com.example.HanldeImg.entity.model.Status;
-import com.example.HanldeImg.repository.ImageRepository;
+import com.example.HanldeImg.scripts.ApkScripts;
 import com.example.HanldeImg.scripts.ImageScripts;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,11 +10,8 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InterruptedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -26,26 +19,32 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ImageServiceImpl implements ImageService {
 
-    private final ImageRepository imageRepository;
     private static final String BASE_UPLOAD_DIRECTORY = "uploads";
 
     private static final Logger log = LoggerFactory.getLogger(ImageServiceImpl.class);
 
+    private static String baseName(String filename) {
+        if (filename == null) return "unknown";
+        String name = filename;
+        int dot = name.lastIndexOf('.');
+        return (dot > 0) ? name.substring(0, dot) : name;
+    }
+
     @Override
     public void uploadProject(List<MultipartFile> files, String projectName) {
-
         Path targetDir = Path.of(BASE_UPLOAD_DIRECTORY, projectName);
 
-
-        for (int i = 0; i < files.size(); i++) {
-            MultipartFile file = files.get(i);
-            if (file.isEmpty()) continue;
-
+        for (MultipartFile file : files) {
+            if (file == null || file.isEmpty()) continue;
 
             try {
-                File tempFile = File.createTempFile("upload_", file.getOriginalFilename());
+                File tempFile = File.createTempFile("upload_", "_" + file.getOriginalFilename());
                 file.transferTo(tempFile);
-                ImageScripts.extractImgWith7z(tempFile.toPath(), targetDir);
+
+                String folderName = baseName(file.getOriginalFilename()); // system / vendor
+                Path imageOutDir = targetDir.resolve(folderName);
+
+                ImageScripts.extractImgWith7z(tempFile.toPath(), imageOutDir);
                 tempFile.delete();
 
             } catch (IOException | InterruptedException e) {
@@ -53,14 +52,12 @@ public class ImageServiceImpl implements ImageService {
             }
         }
 
+        ApkScripts.decodeApksToProjectRoot(targetDir);
 
-        try{
-            ImageScripts.ensure(
-                    System.getenv("GIT_GROUP_NAME"),
-                    projectName
-            );
+        try {
+            ImageScripts.ensure(System.getenv("GIT_GROUP_NAME"), projectName);
             ImageScripts.pushToGitLab(targetDir);
-        }catch (Exception e){
+        } catch (Exception e) {
             log.info("Ошибка при загрузки образа на gitlab");
             throw new RuntimeException(e.getMessage());
         }
